edit FilmAPI
-убрал renderModal с класса FilmAPI
-исправил опечатку в модалке
