+++ src/index.js
@@ -8,4 +8,5 @@ import './js/headerSwitcher';
 import './js/search-film';
 import './js/localStorageMovies';
 import './js/themeswitcher';
+import './js/renderModal';
