+++ src/js/filmAPI.js
@@ -1,21 +1,16 @@
 import axios from 'axios'
 import Notiflix from 'notiflix';
 
-import getRefs from './get-refs';
-import detailFilm from '../templates/detail-film.hbs';
-import LocalStorage from './localStorageMovies'
-import {updateInfoFilms, updateDelInfoFilms, disabledBtn} from './updateInfoFilms'
+import renderModal from './renderModal';
 
 const KEY = '64d8aa762e5eca1f8be6b3971b76ddad'
 const URL = 'https://api.themoviedb.org/3'
 
-const refs = getRefs();
-const localStg = new LocalStorage;
-
 export default class FilmAPI {
     constructor() {
-        this.searchQuery = 'титаник'
-        this.page = 1
+        this.searchQuery = 'титаник';
+        this.page = 1;
+        this.renderModal = renderModal;
     }
 
     get query() {
@@ -28,6 +23,7 @@ export default class FilmAPI {
 
     async searchTrendings() {
         return await axios.get(`${URL}/trending/all/day?api_key=${KEY}&page=${this.page}`)
+        //return await axios.get(`${URL}/discover/movie?api_key=${KEY}&include_adult=false&page=${this.page}`)
     }
 
     async movieById(movieId, filmDataLS) {
@@ -37,10 +33,6 @@ export default class FilmAPI {
         }
         catch (err) {
             this.renderModal(err, filmDataLS)
-            console.error("Error response:");
-            console.error(err.response.data);    // ***
-            console.error(err.response.status);  // ***
-            console.error(err.response.headers); // ***
             Notiflix.Loading.remove();
         }
     }
@@ -60,26 +52,6 @@ export default class FilmAPI {
 
     resetPage() {
         this.page = 1
-    }
-
-    renderModal(film, filmDataLS) {
-        let data = film.data ? film.data : film;
-        const err = data.response;
-        refs.infoFilmIsOpen.classList.toggle('backdrop--is-hidden');
-        refs.bodyEl.classList.toggle('toggle_scroll');
-        if (err) {
-            data = updateDelInfoFilms(JSON.parse(filmDataLS))
-        }
-        data = updateInfoFilms(data)
-        refs.infoFilmContainer.insertAdjacentHTML('beforeend', detailFilm(data));
-        if (err) {
-            disabledBtn()
-        }
-        if (!refs.infoFilmIsOpen.classList.contains("backdrop--is-hidden") && !err) {
-            localStg.changeDataBtn(data.id);
-            localStg.eventWatchedQueueBtn(filmDataLS)
-        }
-        Notiflix.Loading.remove();
     }
 }
